fix(YAGHTM-3): Clean up code
