--- yaghm/plugin/src/main/kotlin/dev/yaghm/plugin/internal/core/dsl/githook/GitHookAction.kt
@@ -1,7 +1,0 @@
-package dev.yaghm.plugin.internal.core.dsl.githook
-
-import dev.yaghm.plugin.internal.config.GitHookConfig
-
-class GitHookAction {
-    var command: String? = null
-}
